refactor(web): remove creator section bg
