+++ apps/web/src/app/about/_components/creator-section.tsx
@@ -1,0 +1,67 @@
+"use client";
+
+import { Link } from "@heroui/link";
+import Typography from "@web/components/typography";
+import { motion, useReducedMotion } from "framer-motion";
+import { Heart } from "lucide-react";
+import { staggerContainerVariants, staggerItemVariants } from "./variants";
+
+export const CreatorSection = () => {
+  const shouldReduceMotion = useReducedMotion();
+
+  return (
+    <section className="-mx-6 relative overflow-hidden px-6 py-20 lg:py-28">
+      {/* Subtle decorative elements */}
+      <div className="pointer-events-none absolute inset-0 overflow-hidden">
+        <div className="absolute top-20 left-10 h-px w-32 bg-gradient-to-r from-transparent via-default-300 to-transparent opacity-50" />
+        <div className="absolute top-40 right-20 h-px w-24 bg-gradient-to-r from-transparent via-default-300 to-transparent opacity-40" />
+        <div className="absolute bottom-32 left-1/4 h-px w-40 bg-gradient-to-r from-transparent via-default-300 to-transparent opacity-30" />
+      </div>
+
+      <div className="container relative mx-auto">
+        <motion.div
+          className="flex flex-col items-center"
+          variants={shouldReduceMotion ? undefined : staggerContainerVariants}
+          initial={shouldReduceMotion ? undefined : "hidden"}
+          whileInView="visible"
+          viewport={{ once: true, amount: 0.2 }}
+        >
+          <motion.div
+            className="mx-auto flex max-w-2xl flex-col items-center gap-8 text-center"
+            variants={shouldReduceMotion ? undefined : staggerItemVariants}
+          >
+            <Typography.Label className="text-primary uppercase tracking-widest">
+              Behind the Data
+            </Typography.Label>
+
+            <div className="flex flex-col gap-5">
+              <Typography.H2 className="text-foreground">
+                Built with{" "}
+                <Heart className="inline size-6 fill-red-500 text-red-500" /> in
+                Singapore
+              </Typography.H2>
+              <Typography.TextLg className="text-foreground/70">
+                SG Cars Trends is an independent project created by{" "}
+                <Link
+                  href="https://ruchern.dev"
+                  isExternal
+                  className="font-medium text-primary transition-colors hover:text-primary/80"
+                >
+                  Ru Chern
+                </Link>
+                , a software engineer who wanted to make this data easier to
+                explore.
+              </Typography.TextLg>
+            </div>
+
+            <Typography.TextSm className="max-w-lg text-foreground/50">
+              This platform is maintained in spare time alongside a full-time
+              job. If you find it useful, consider sharing it with others who
+              might benefit.
+            </Typography.TextSm>
+          </motion.div>
+        </motion.div>
+      </div>
+    </section>
+  );
+};
