+++ apps/web/src/app/about/page.tsx
@@ -2,7 +2,8 @@ import { StructuredData } from "@web/components/structured-data";
 import { SITE_TITLE, SITE_URL } from "@web/config";
 import type { Metadata } from "next";
 import type { Organization, Person, WebPage, WithContext } from "schema-dts";
-import { CommunitySection } from "./_components/community-section";
+import { CreatorSection } from "./_components/creator-section";
+import { CtaSection } from "./_components/cta-section";
 import { DataSection } from "./_components/data-section";
 import { HeroSection } from "./_components/hero-section";
 import { MissionSection } from "./_components/mission-section";
@@ -102,7 +103,8 @@ const AboutPage = () => {
         <MissionSection />
         <DataSection />
         <TimelineSection />
-        <CommunitySection />
+        <CreatorSection />
+        <CtaSection />
       </div>
     </>
   );
